Move geometry_tree.py
