+++ src/ansys/motorcad/core/methods/adaptive_geometry.py
@@ -24,7 +24,7 @@
 from warnings import warn
 
 from ansys.motorcad.core.geometry import Region, RegionMagnet
-from ansys.motorcad.core.methods.geometry_tree import GeometryTree
+from ansys.motorcad.core.geometry_tree import GeometryTree
 from ansys.motorcad.core.rpc_client_core import MotorCADError, is_running_in_internal_scripting
